requirements.txt
easydict==1.6
libvirt-python==1.3.2
pexpect==4.0.1
ptyprocess==0.5.1
